Auto delete rejected documents

+++ loader/src/main/java/com/receiptofi/loader/scheduledtasks/FileSystemProcess.java
@@ -1,4 +1,4 @@
-package com.receiptofi.web.scheduledtasks;
+package com.receiptofi.loader.scheduledtasks;
 
 import com.receiptofi.service.ReceiptService;
 import com.receiptofi.utils.CreateTempFile;
@@ -37,27 +37,35 @@ public class FileSystemProcess {
     @Value("${deleteExcelFileAfterDay:7}")
     private int deleteExcelFileAfterDay;
 
+    //TODO(hth) add to AOP to turn on and off instead
+    @Value("${removeExpiredExcelFiles:ON}")
+    private String removeExpiredExcelFiles;
+
     @Autowired private ReceiptService receiptService;
 
     //for every two second use */2 * * * * ? where as cron string blow run every day at 12:00 AM
     @Scheduled(cron="0 0 0 * * ?")
     public void removeExpiredExcelFiles() {
-        LOG.info("begins");
-        int count = 0, found = 0;
-        try {
-            AgeFileFilter cutoff = new AgeFileFilter(DateUtil.now().minusDays(deleteExcelFileAfterDay).toDate());
-            File directory = new File(expensofiReportLocation);
-            String[] files = directory.list(cutoff);
-            found = files.length;
-            for(String filename : files) {
-                removeExpiredExcel(getExcelFile(filename));
-                receiptService.removeExpensofiFilenameReference(filename);
-                count++;
+        if(removeExpiredExcelFiles.equalsIgnoreCase("ON")) {
+            LOG.info("feature is {}", removeExpiredExcelFiles);
+            int count = 0, found = 0;
+            try {
+                AgeFileFilter cutoff = new AgeFileFilter(DateUtil.now().minusDays(deleteExcelFileAfterDay).toDate());
+                File directory = new File(expensofiReportLocation);
+                String[] files = directory.list(cutoff);
+                found = files.length;
+                for (String filename : files) {
+                    removeExpiredExcel(getExcelFile(filename));
+                    receiptService.removeExpensofiFilenameReference(filename);
+                    count++;
+                }
+            } catch (Exception e) {
+                LOG.error("found error={}", e.getLocalizedMessage(), e);
+            } finally {
+                LOG.info("complete deletedExcelFile={}, foundExcelFile={}", count, found);
             }
-        } catch(Exception e) {
-            LOG.error("found error={}", e.getLocalizedMessage(), e);
-        }finally {
-            LOG.info("delete complete deletedExcelFile={}, foundExcelFile={}", count, found);
+        } else {
+            LOG.info("feature is {}", removeExpiredExcelFiles);
         }
     }
